Base for Colorblind Mode
WARNING: Currently crashes the game if Colorblind Mode is on with the latest commit pulled from the asset repo.

Once fixed, this shall be backported to the master branch, and eventually to the main repository.

Colorblind Mode can be turned on by setting "colorblindMode" to true in the "options" field in runtime.json.