Incomplete time-dynamic music
This has to be revamped. Right now it's only using the more advanced portions of the tracks.

Zuma Blitz's ingame tracks progress as you play longer, including for how long you've been in the 15-minute danger mark.

The music one-offs for the time running out and the game launching will probably be done soon.