Brendan's hypothesis for gap shot scoring
"For gap shots I don't have any exact calculations for Kroakatoa since there was a ton of balancing regarding this, and it's almost impossible to know the formula without seeing the source code because (I think) it is the sum of an inverse and exponential function, with a max of 11000 at 0 ball widths, and a minimum of 500 points at around 20 ball widths. 0 ball gap shots are not possible here since the game calculates the gap size when the match is made, instead of when the ball was fired."

Looking at the source code would be near-impossible due to the fact that the Kroakatoa main SWF file is obfuscated - thus I'm settling for an estimated calculation.

(Formula figured out by @cheffplayer)

Co-Authored-By: cheffplayer <9261929+cheffplayer@users.noreply.github.com>
Co-Authored-By: Brendan Chan <b_chan1@u.pacific.edu>