Prevent spheres from rolling back if magnetization is blocked
Adds the following check:
- Is the gap size greater than zero?
- Is the next group's first sphere color and the current group's last sphere color the same?
If both of these are yes, stop the next group from rolling.

This currently does not account for scarabs - this must be refactored before implementing to OpenSMCE upstream.

It seems like isMagnetizing still returns true even after an attempt has been made to block magnetization, so this check is possible.

Fixes #14